Refactor: v5 use the new ts-generator factories
