+++ src/clients/TokenLockingClient.ts
@@ -2,7 +2,7 @@ import { Signer } from 'ethers';
 import { Provider } from 'ethers/providers';
 
 import { ClientType } from '../constants';
-import { TokenLockingFactory } from '../contracts/5/TokenLockingFactory';
+import { TokenLocking__factory as TokenLockingFactory } from '../contracts/5/factories/TokenLocking__factory';
 import { TokenLocking } from '../contracts/5/TokenLocking';
 
 export interface TokenLockingClient extends TokenLocking {
